+++ _examples/view/quicktemplate/controllers/hello.go
@@ -1,7 +1,7 @@
 package controllers
 
 import (
-	"github.com/kataras/iris/v12/_examples/response-writer/quicktemplate/templates"
+	"github.com/kataras/iris/v12/_examples/view/quicktemplate/templates"
 
 	"github.com/kataras/iris/v12"
 )